+++ src/NetAppFiles/NetAppFiles/Helpers/ModelExtensions.cs
@@ -226,7 +226,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Helpers
 
         public static PSNetAppFilesVolumeDataProtection ConvertDataProtectionToPs(VolumePropertiesDataProtection DataProtection)
         {
-            var psDataProtection = new PSNetAppFilesVolumeDataProtection();            
+            var psDataProtection = new PSNetAppFilesVolumeDataProtection();
             if (DataProtection.Replication != null)
             {
                 var replication = new PSNetAppFilesReplicationObject();
@@ -248,7 +248,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Helpers
             if (DataProtection.Backup != null)
             {
                 var psBackupProps = new PSNetAppFilesVolumeBackupProperties()
-                {                    
+                {
                     BackupPolicyId = DataProtection.Backup.BackupPolicyId,
                     PolicyEnforced = DataProtection.Backup.PolicyEnforced,
                     BackupVaultId = DataProtection.Backup.BackupVaultId
@@ -260,8 +260,14 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Helpers
                 var volumeRelocation = DataProtection.VolumeRelocation.ConvertToPs();
                 psDataProtection.VolumeRelocation = volumeRelocation;
             }
-           return psDataProtection;
+            if (DataProtection.RansomwareProtection != null)
+            {
+                var ransomwareProtection = DataProtection.RansomwareProtection.ConvertToPs();
+                psDataProtection.RansomwareProtection = ransomwareProtection;
+            }
+            return psDataProtection;
         }
+
         public static PSRemotePath ConvertToPs(this RemotePath remotePath)
         {
             var psRemotePath = new PSRemotePath();
@@ -313,6 +319,12 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Helpers
                 backup.PolicyEnforced = psDataProtection.Backup.PolicyEnforced;
                 backup.BackupVaultId = psDataProtection.Backup.BackupVaultId;
                 dataProtection.Backup = backup;
+            }
+            if (psDataProtection.RansomwareProtection != null)
+            {
+                var ransomwareProtection = new RansomwareProtectionSettings();
+                ransomwareProtection.DesiredRansomwareProtectionState = psDataProtection.RansomwareProtection.DesiredRansomwareProtectionState;
+                dataProtection.RansomwareProtection = ransomwareProtection;
             }
             return dataProtection;
         }
@@ -402,6 +414,12 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Helpers
                 backup.BackupVaultId = psDataProtection.Backup.BackupVaultId;
                 dataProtection.Backup = backup;
             }
+            if (psDataProtection.RansomwareProtection != null)
+            {
+                var ransomwareProtection = new RansomwareProtectionPatchSettings();
+                ransomwareProtection.DesiredRansomwareProtectionState = psDataProtection.RansomwareProtection.DesiredRansomwareProtectionState;
+                dataProtection.RansomwareProtection = ransomwareProtection;
+            }
             return dataProtection;
         }
 
@@ -449,6 +467,14 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Helpers
             psVolumeRelocation.RelocationRequested = volumeRelocation.RelocationRequested;
             psVolumeRelocation.ReadyToBeFinalized = volumeRelocation.ReadyToBeFinalized;
             return psVolumeRelocation;
+        }
+
+        public static PSNetAppFilesVolumeRansomwareProperties ConvertToPs(this RansomwareProtectionSettings ransomwareProtection)
+        {
+            var psRansomwareProtection = new PSNetAppFilesVolumeRansomwareProperties();
+            psRansomwareProtection.DesiredRansomwareProtectionState = ransomwareProtection.DesiredRansomwareProtectionState;
+            psRansomwareProtection.ActualRansomwareProtectionState = ransomwareProtection.ActualRansomwareProtectionState;
+            return psRansomwareProtection;
         }
 
         public static PSNetAppFilesVolume ToPsNetAppFilesVolume(this Management.NetApp.Models.Volume volume)